requirements.txt
openai
questionary
rich
pyyaml
